Give the histogram and the colorspace panel their own files
image-gui.cpp held three things that never call each other: the histogram
and its helpers, the CIE diagram with the colorspace panel, and the info
panel with the layer tree and the channel statistics. They are three
files now, the way app-*.cpp is already split by concern.

Nothing else changed. Each new file's body is a contiguous slice of the
old one, byte for byte, and the three share no helper, so no static had to
open up. The includes went with the code that needed them; shader.h went
nowhere, since none of the three uses it.

319 cases, 359,225 assertions; hdrview_gui_tests 101/101.
